@@ -29,72 +29,72 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   final audios = <Audio>[
     Audio.network(
-      "https://files.freemusicarchive.org/storage-freemusicarchive-org/music/Music_for_Video/springtide/Sounds_strange_weird_but_unmistakably_romantic_Vol1/springtide_-_03_-_We_Are_Heading_to_the_East.mp3",
+      'https://files.freemusicarchive.org/storage-freemusicarchive-org/music/Music_for_Video/springtide/Sounds_strange_weird_but_unmistakably_romantic_Vol1/springtide_-_03_-_We_Are_Heading_to_the_East.mp3',
       metas: Metas(
-        title: "Online",
-        artist: "Florent Champigny",
-        album: "OnlineAlbum",
+        title: 'Online',
+        artist: 'Florent Champigny',
+        album: 'OnlineAlbum',
         image: MetasImage.network(
-            "https://image.shutterstock.com/image-vector/pop-music-text-art-colorful-600w-515538502.jpg"),
+            'https://image.shutterstock.com/image-vector/pop-music-text-art-colorful-600w-515538502.jpg'),
       ),
     ),
     Audio(
-      "assets/audios/rock.mp3",
+      'assets/audios/rock.mp3',
       metas: Metas(
-        title: "Rock",
-        artist: "Florent Champigny",
-        album: "RockAlbum",
+        title: 'Rock',
+        artist: 'Florent Champigny',
+        album: 'RockAlbum',
         image: MetasImage.network(
-            "https://static.radio.fr/images/broadcasts/cb/ef/2075/c300.png"),
+            'https://static.radio.fr/images/broadcasts/cb/ef/2075/c300.png'),
       ),
     ),
     Audio(
-      "assets/audios/country.mp3",
+      'assets/audios/country.mp3',
       metas: Metas(
-        title: "Country",
-        artist: "Florent Champigny",
-        album: "CountryAlbum",
-        image: MetasImage.asset("assets/images/country.jpg"),
+        title: 'Country',
+        artist: 'Florent Champigny',
+        album: 'CountryAlbum',
+        image: MetasImage.asset('assets/images/country.jpg'),
       ),
     ),
     Audio(
-      "assets/audios/electronic.mp3",
+      'assets/audios/electronic.mp3',
       metas: Metas(
-        title: "Electronic",
-        artist: "Florent Champigny",
-        album: "ElectronicAlbum",
+        title: 'Electronic',
+        artist: 'Florent Champigny',
+        album: 'ElectronicAlbum',
         image: MetasImage.network(
-            "https://i.ytimg.com/vi/nVZNy0ybegI/maxresdefault.jpg"),
+            'https://99designs-blog.imgix.net/blog/wp-content/uploads/2017/12/attachment_68585523.jpg'),
       ),
     ),
     Audio(
-      "assets/audios/hiphop.mp3",
+      'assets/audios/hiphop.mp3',
       metas: Metas(
-        title: "HipHop",
-        artist: "Florent Champigny",
-        album: "HipHopAlbum",
+        title: 'HipHop',
+        artist: 'Florent Champigny',
+        album: 'HipHopAlbum',
         image: MetasImage.network(
-            "https://beyoudancestudio.ch/wp-content/uploads/2019/01/apprendre-danser.hiphop-1.jpg"),
+            'https://beyoudancestudio.ch/wp-content/uploads/2019/01/apprendre-danser.hiphop-1.jpg'),
       ),
     ),
     Audio(
-      "assets/audios/pop.mp3",
+      'assets/audios/pop.mp3',
       metas: Metas(
-        title: "Pop",
-        artist: "Florent Champigny",
-        album: "PopAlbum",
+        title: 'Pop',
+        artist: 'Florent Champigny',
+        album: 'PopAlbum',
         image: MetasImage.network(
-            "https://image.shutterstock.com/image-vector/pop-music-text-art-colorful-600w-515538502.jpg"),
+            'https://image.shutterstock.com/image-vector/pop-music-text-art-colorful-600w-515538502.jpg'),
       ),
     ),
     Audio(
-      "assets/audios/instrumental.mp3",
+      'assets/audios/instrumental.mp3',
       metas: Metas(
-        title: "Instrumental",
-        artist: "Florent Champigny",
-        album: "InstrumentalAlbum",
+        title: 'Instrumental',
+        artist: 'Florent Champigny',
+        album: 'InstrumentalAlbum',
         image: MetasImage.network(
-            "https://i.ytimg.com/vi/zv_0dSfknBc/maxresdefault.jpg"),
+            'https://99designs-blog.imgix.net/blog/wp-content/uploads/2017/12/attachment_68585523.jpg'),
       ),
     ),
   ];
@@ -105,16 +105,16 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     _subscriptions.add(_assetsAudioPlayer.playlistFinished.listen((data) {
-      print("finished : $data");
+      print('finished : $data');
     }));
     _subscriptions.add(_assetsAudioPlayer.playlistAudioFinished.listen((data) {
-      print("playlistAudioFinished : $data");
+      print('playlistAudioFinished : $data');
     }));
     _subscriptions.add(_assetsAudioPlayer.current.listen((data) {
-      print("current : $data");
+      print('current : $data');
     }));
     _subscriptions.add(_assetsAudioPlayer.onReadyToPlay.listen((audio) {
-      print("onRedayToPlay : $audio");
+      print('onRedayToPlay : $audio');
     }));
     super.initState();
   }
@@ -150,38 +150,37 @@ class _MyAppState extends State<MyApp> {
                     StreamBuilder(
                       stream: _assetsAudioPlayer.current,
                       builder: (BuildContext context,
-                          AsyncSnapshot<Playing> snapshot) {
-                        final Playing playing = snapshot.data;
-                        if (playing != null) {
-                          final myAudio =
-                              find(this.audios, playing.audio.assetAudioPath);
-                          return Padding(
-                            padding: const EdgeInsets.all(8.0),
-                            child: Neumorphic(
-                              style: NeumorphicStyle(
-                                depth: 8,
-                                surfaceIntensity: 1,
-                                shape: NeumorphicShape.concave,
-                                boxShape: NeumorphicBoxShape.circle(),
-                              ),
-                              child:
-                                  myAudio.metas.image.type == ImageType.network
-                                      ? Image.network(
-                                          myAudio.metas.image.path,
-                                          height: 150,
-                                          width: 150,
-                                          fit: BoxFit.contain,
-                                        )
-                                      : Image.asset(
-                                          myAudio.metas.image.path,
-                                          height: 150,
-                                          width: 150,
-                                          fit: BoxFit.contain,
-                                        ),
+                          AsyncSnapshot<Playing?> snapshot) {
+                        if (!snapshot.hasData) return const SizedBox();
+                        final playing = snapshot.data!;
+                        final myAudio =
+                            find(audios, playing.audio.assetAudioPath);
+                        return Padding(
+                          padding: const EdgeInsets.all(8.0),
+                          child: Neumorphic(
+                            style: NeumorphicStyle(
+                              depth: 8,
+                              surfaceIntensity: 1,
+                              shape: NeumorphicShape.concave,
+                              boxShape: NeumorphicBoxShape.circle(),
                             ),
-                          );
-                        }
-                        return SizedBox();
+                            child: myAudio.metas.image?.path == null
+                                ? const SizedBox()
+                                : myAudio.metas.image?.type == ImageType.network
+                                    ? Image.network(
+                                        myAudio.metas.image!.path,
+                                        height: 150,
+                                        width: 150,
+                                        fit: BoxFit.contain,
+                                      )
+                                    : Image.asset(
+                                        myAudio.metas.image!.path,
+                                        height: 150,
+                                        width: 150,
+                                        fit: BoxFit.contain,
+                                      ),
+                          ),
+                        );
                       },
                     ),
                     Align(
@@ -194,7 +193,7 @@ class _MyAppState extends State<MyApp> {
                         margin: EdgeInsets.all(18),
                         onPressed: () {
                           AssetsAudioPlayer.playAndForget(
-                              Audio("assets/audios/horn.mp3"));
+                              Audio('assets/audios/horn.mp3'));
                         },
                         child: Icon(
                           Icons.add_alert,
@@ -212,23 +211,30 @@ class _MyAppState extends State<MyApp> {
                 ),
                 StreamBuilder(
                     stream: _assetsAudioPlayer.current,
-                    builder: (context, snapshot) {
-                      if (!snapshot.hasData) {
-                        return SizedBox();
-                      }
-                      final Playing playing = snapshot.data;
+                    builder: (context, AsyncSnapshot<Playing?> snapshot) {
+                      if (!snapshot.hasData) return const SizedBox();
+
+                      final playing = snapshot.data!;
                       return Column(
                         children: <Widget>[
                           StreamBuilder(
                             stream: _assetsAudioPlayer.loopMode,
                             initialData: LoopMode.none,
-                            builder: (context, snapshotLooping) {
-                              final LoopMode loopMode = snapshotLooping.data;
+                            builder: (context,
+                                AsyncSnapshot<LoopMode> snapshotLooping) {
+                              if (!snapshotLooping.hasData) {
+                                return const SizedBox();
+                              }
+                              final loopMode = snapshotLooping.data!;
                               return StreamBuilder(
                                   stream: _assetsAudioPlayer.isPlaying,
                                   initialData: false,
-                                  builder: (context, snapshotPlaying) {
-                                    final isPlaying = snapshotPlaying.data;
+                                  builder: (context,
+                                      AsyncSnapshot<bool> snapshotPlaying) {
+                                    if (!snapshotPlaying.hasData) {
+                                      return const SizedBox();
+                                    }
+                                    final isPlaying = snapshotPlaying.data!;
                                     return PlayingControls(
                                       loopMode: loopMode,
                                       isPlaying: isPlaying,
@@ -253,13 +259,14 @@ class _MyAppState extends State<MyApp> {
                           ),
                           StreamBuilder(
                               stream: _assetsAudioPlayer.realtimePlayingInfos,
-                              builder: (context, snapshot) {
+                              builder: (context,
+                                  AsyncSnapshot<RealtimePlayingInfos>
+                                      snapshot) {
                                 if (!snapshot.hasData) {
-                                  return SizedBox();
+                                  return const SizedBox();
                                 }
-                                final RealtimePlayingInfos infos =
-                                    snapshot.data;
-                                //print("infos: $infos");
+                                final infos = snapshot.data!;
+                                // print('infos: $infos');
                                 return PositionSeekWidget(
                                   currentPosition: infos.currentPosition,
                                   duration: infos.duration,
@@ -278,10 +285,11 @@ class _MyAppState extends State<MyApp> {
                   child: StreamBuilder(
                       stream: _assetsAudioPlayer.current,
                       builder: (BuildContext context,
-                          AsyncSnapshot<Playing> snapshot) {
-                        final Playing playing = snapshot.data;
+                          AsyncSnapshot<Playing?> snapshot) {
+                        if (!snapshot.hasData) return const SizedBox();
+                        final playing = snapshot.data!;
                         return SongsSelector(
-                          audios: this.audios,
+                          audios: audios,
                           onPlaylistSelected: (myAudios) {
                             _assetsAudioPlayer.open(
                               Playlist(audios: myAudios),
@@ -303,8 +311,9 @@ class _MyAppState extends State<MyApp> {
                 StreamBuilder(
                     stream: _assetsAudioPlayer.volume,
                     initialData: AssetsAudioPlayer.defaultVolume,
-                    builder: (context, snapshot) {
-                      final double volume = snapshot.data;
+                    builder: (context, AsyncSnapshot<double> snapshot) {
+                      if (!snapshot.hasData) return const SizedBox();
+                      final volume = snapshot.data!;
                       return VolumeSelector(
                         volume: volume,
                         onChange: (v) {
@@ -315,8 +324,9 @@ class _MyAppState extends State<MyApp> {
                 StreamBuilder(
                     stream: _assetsAudioPlayer.forwardRewindSpeed,
                     initialData: null,
-                    builder: (context, snapshot) {
-                      final double speed = snapshot.data;
+                    builder: (context, AsyncSnapshot<double?> snapshot) {
+                      if (!snapshot.hasData) return const SizedBox();
+                      final speed = snapshot.data!;
                       return ForwardRewindSelector(
                         speed: speed,
                         onChange: (v) {
@@ -327,8 +337,9 @@ class _MyAppState extends State<MyApp> {
                 StreamBuilder(
                     stream: _assetsAudioPlayer.playSpeed,
                     initialData: AssetsAudioPlayer.defaultPlaySpeed,
-                    builder: (context, snapshot) {
-                      final double playSpeed = snapshot.data;
+                    builder: (context, AsyncSnapshot<double> snapshot) {
+                      if (!snapshot.hasData) return const SizedBox();
+                      final playSpeed = snapshot.data!;
                       return PlaySpeedSelector(
                         playSpeed: playSpeed,
                         onChange: (v) {
